BG Remover/requirements.txt
pillow==10.4.0
rembg==2.0.59
